LLM/hanhs-on-llm/chapter-04/walk_through/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cryptography==46.0.6
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
emoji==2.15.0
et_xmlfile==2.0.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-vision==3.13.0
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
installer==0.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.22
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
olefile==0.47
openai-whisper==20250625
openpyxl==3.1.5
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.3.0
pikepdf==10.5.1
pillow==12.1.1
preshed==3.0.13
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2.2
pypdf==6.9.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==1.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.22.6
unstructured-client==0.42.12
unstructured.pytesseract==0.3.15
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
weasel==1.0.0
webencodings==0.5.1
wrapt==2.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
